test-requirements.txt
# Linting
flake8
pycodestyle

# Unit testing
mock
pytest
pytest-cov

# Integration testing
behave
PyHamcrest
